Add NodeList custom list for nodes with list filtering.
NOTE: Backwards incompatible change due to removal of the now-unnecessary
      `children_in_ns`, `get` and `get_first` methods.

Node lists wrapped by the new NodeList class can be filtered by invoking
the `filter` method (or the list as a callable) and supplying filtering
criteria, or an arbitrary filter function. The first filtered result (or None)
is also directly available through the `first` method.