Add `child()` method to Node to get first child with filtering.
BACKWARDS-INCOMPATIBLE API CHANGE

The new `child()` method replaces the technique of calling the `first()`
method on a NodeList, e.g. as returned by the `children` attribute.
There remains a `first` attribute on NodeList to return the first of
a set of filtered children, but the `child()` method will be the
preferred way of doing this .