Refactor node import to explicit transplant/clone methods
BACKWARDS INCOMPATIBLE API CHANGE

The prior implementation for importing nodes from one document into
another suffered from ambiguous and different behaviour depending on the
underlying XML library implementation. Specifically, an import-with-copy
was broken for the minidom adapter.

To fix this, and to remove the ambiguity about when an imported node
would or would not be removed from the source document, there are now
two explicit operations:
 - clone node: copies the node and its descendants into the current
   document without removing it from the source document.
 - transplant node: moves the node and its descendants into the current
   document and removes it from the source document.

Where necessary, the adapter implementations will now do any extra work
required to provide this explicit and consistent behaviour.