Make separate adapters for pure-Python and C-based ElementTree library.
Make it possible to create an xml4h adapter for either the C or non-C
ElementTree implementation, rather than assuming that the C version
should be used automatically and always when it is available.

This will allow users with existing non-C-based ElementTree-based code,
or those who prefer not to use the C-based version for some reason,
to take advantage of xml4h without weird incompatibility errors.