Fix subtle bug replacing attributes with `attributes` property setter
Fixed a subtle issue when replacing all of an element's attributes by
assigning a new attributes dictionary via the `element.attributes`
property setter, which affected at least the ElementTree implementation
with Python 2.7 (though seemingly not with Python 2.6)

If an 'xmlns' namespacing attribute was set, and another attribute on
the element used this namespace, and the 'xmlns' attribute happened to
be removed *before* the other attribute that relied on that namespace,
then the attribute with the now-defunct namespace would not be removed
properly.

This bug would cause the bad Python 2.7 tests to fail intermittently,
depending on the undefined order in which attributes happened to be
traversed for removal.